requirements.txt
cloudinary
dotenv
asgiref==3.7.2
certifi==2023.7.22
cryptography==41.0.3
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
psycopg2-binary
PyJWT==2.8.0
requests==2.31.0
requests-oauthlib==1.3.1
